Update CTPanoramaView.swift
Making changes this way eliminates the need to make messy camera adjustments and prevents the compass from starting at a weird angle. You're basically establishing which way is forward. There could be room to use a quaternion here rather than an angle. I used a float since it better supports cylindrical panoramas and maintains a simpler interface. It would be easy to swap it out though